refactor interrupts
